refactor(database): simplify one-to-x relation to not require "relationId" columns refactor(database): remove relation renames in select refactor(database): propagate embedded schema logic to postgres refactor(database): re-merge sql handling refactor(authentication): socialDetails fields schema
