+++ modules/database/src/adapters/sequelize-adapter/index.ts
@@ -15,15 +15,23 @@ import { status } from '@grpc/grpc-js';
 import { SequelizeAuto } from 'sequelize-auto';
 import { DatabaseAdapter } from '../DatabaseAdapter';
 import { SequelizeSchema } from './SequelizeSchema';
-import { checkIfPostgresOptions, tableFetch } from './utils';
+import {
+  checkIfPostgresOptions,
+  compileSchema,
+  resolveRelatedSchemas,
+  tableFetch,
+} from './utils';
 import { sqlSchemaConverter } from '../../introspection/sequelize/utils';
-import { introspectedSchemaCmsOptionsDefaults } from '../../interfaces';
+import {
+  ConduitDatabaseSchema,
+  introspectedSchemaCmsOptionsDefaults,
+} from '../../interfaces';
+import { schemaConverter } from './sql-adapter/SchemaConverter';
+import { isNil } from 'lodash';
 
 const sqlSchemaName = process.env.SQL_SCHEMA ?? 'public';
 
-export abstract class SequelizeAdapter<
-  T extends SequelizeSchema,
-> extends DatabaseAdapter<T> {
+export abstract class SequelizeAdapter extends DatabaseAdapter<SequelizeSchema> {
   connectionUri: string;
   sequelize!: Sequelize;
   readonly SUPPORTED_DIALECTS = ['postgres', 'mysql', 'sqlite', 'mariadb'];
@@ -161,13 +169,135 @@ export abstract class SequelizeAdapter<
       : schema.name;
   }
 
-  protected abstract _createSchemaFromAdapter(
-    schema: ConduitSchema,
-    saveToDb: boolean,
+  private async processExtractedSchemas(
+    schema: ConduitDatabaseSchema,
+    extractedSchemas: Indexable,
+    associatedSchemas: { [key: string]: SequelizeSchema | SequelizeSchema[] },
+  ) {
+    for (const extractedSchema in extractedSchemas) {
+      const modelOptions = {
+        ...schema.modelOptions,
+        permissions: {
+          extendable: false,
+          canCreate: false,
+          canModify: 'Nothing',
+          canDelete: false,
+        },
+      };
+      let modeledSchema;
+      let isArray = false;
+      if (Array.isArray(extractedSchemas[extractedSchema])) {
+        isArray = true;
+        modeledSchema = new ConduitSchema(
+          `${schema.name}_${extractedSchema}`,
+          extractedSchemas[extractedSchema][0],
+          modelOptions,
+          `${schema.collectionName}_${extractedSchema}`,
+        );
+      } else {
+        modeledSchema = new ConduitSchema(
+          `${schema.name}_${extractedSchema}`,
+          extractedSchemas[extractedSchema],
+          modelOptions,
+          `${schema.collectionName}_${extractedSchema}`,
+        );
+      }
+
+      modeledSchema.ownerModule = schema.ownerModule;
+      (modeledSchema as ConduitDatabaseSchema).compiledFields = modeledSchema.fields;
+      // check index compatibility
+      const sequelizeSchema = await this._createSchemaFromAdapter(
+        modeledSchema as ConduitDatabaseSchema,
+        false,
+        {
+          parentSchema: schema.name,
+        },
+      );
+      associatedSchemas[extractedSchema] = isArray ? [sequelizeSchema] : sequelizeSchema;
+    }
+  }
+
+  protected async _createSchemaFromAdapter(
+    schema: ConduitDatabaseSchema,
+    saveToDb: boolean = true,
     options?: { parentSchema: string },
-  ): Promise<T>;
+  ): Promise<SequelizeSchema> {
+    const compiledSchema = compileSchema(
+      schema,
+      this.registeredSchemas,
+      this.sequelize.models,
+    );
+
+    const [newSchema, extractedSchemas, extractedRelations] =
+      schemaConverter(compiledSchema);
+    this.registeredSchemas.set(
+      schema.name,
+      Object.freeze(JSON.parse(JSON.stringify(schema))),
+    );
+    const relatedSchemas = await resolveRelatedSchemas(
+      schema,
+      extractedRelations,
+      this.models,
+    );
+    const associatedSchemas: { [key: string]: SequelizeSchema | SequelizeSchema[] } = {};
+    await this.processExtractedSchemas(schema, extractedSchemas, associatedSchemas);
+    if (options && options.parentSchema) {
+      schema.parentSchema = options.parentSchema;
+    }
+    this.models[schema.name] = new SequelizeSchema(
+      this.sequelize,
+      newSchema,
+      schema,
+      this,
+      relatedSchemas,
+      associatedSchemas,
+    );
+
+    const noSync = this.models[schema.name].originalSchema.modelOptions.conduit!.noSync;
+    // do not sync extracted schemas
+    if ((isNil(noSync) || !noSync) && !options) {
+      await this.models[schema.name].sync();
+    }
+    // do not store extracted schemas to db
+    if (!options && saveToDb) {
+      await this.compareAndStoreMigratedSchema(schema);
+      await this.saveSchemaToDatabase(schema);
+      if (associatedSchemas && Object.keys(associatedSchemas).length > 0) {
+        for (const associatedSchema in associatedSchemas) {
+          const schema = associatedSchemas[associatedSchema];
+          if (Array.isArray(schema)) {
+            await this.saveSchemaToDatabase(schema[0].originalSchema);
+          } else {
+            await this.saveSchemaToDatabase(schema.originalSchema);
+          }
+        }
+      }
+    }
+    return this.models[schema.name];
+  }
 
   async deleteSchema(
+    schemaName: string,
+    deleteData: boolean,
+    callerModule: string = 'database',
+    instanceSync = false,
+  ): Promise<string> {
+    if ((instanceSync || deleteData) && this.models[schemaName].associations) {
+      for (const association in this.models[schemaName].associations) {
+        const associationSchema = this.models[schemaName].associations![association];
+        if (Array.isArray(associationSchema)) {
+          delete this.models[associationSchema[0].schema.name];
+          if (deleteData) await associationSchema[0].model.drop();
+        } else {
+          delete this.models[associationSchema.schema.name];
+          if (deleteData) await associationSchema.model.drop();
+        }
+      }
+    }
+    return await this._deleteSchema(schemaName, deleteData, callerModule, instanceSync);
+  }
+
+  private async _deleteSchema(
     schemaName: string,
     deleteData: boolean,
     callerModule: string = 'database',
@@ -206,14 +336,14 @@ export abstract class SequelizeAdapter<
     return 'Schema deleted!';
   }
 
-  getSchemaModel(schemaName: string): { model: T } {
+  getSchemaModel(schemaName: string): { model: SequelizeSchema } {
     if (this.models && this.models[schemaName]) {
       return { model: this.models[schemaName] };
     }
     throw new GrpcError(status.NOT_FOUND, `Schema ${schemaName} not defined yet`);
   }
 
-  // change to dialect
+  //todo change to dialect
   getDatabaseType(): string {
     return 'PostgreSQL';
   }
