+++ modules/database/src/adapters/sequelize-adapter/utils/extractors/index.ts
@@ -1,7 +1,32 @@
+import { ConduitModel, Indexable } from '@conduitplatform/grpc-sdk';
 import { isArray, isObject } from 'lodash';
-import { ConduitModel } from '@conduitplatform/grpc-sdk';
 
-interface RelationType {
+export function extractEmbedded(ogSchema: any, schema: any) {
+  const extracted: Indexable = {};
+  for (const key of Object.keys(schema)) {
+    if (isArray(schema[key])) {
+      const arrayField = schema[key];
+      if (arrayField[0] !== null && typeof arrayField[0] === 'object') {
+        if (
+          !arrayField[0].hasOwnProperty('type') ||
+          typeof arrayField[0].type !== 'string'
+        ) {
+          extracted[key] = [arrayField[0]];
+          delete schema[key];
+          delete ogSchema[key];
+        }
+      }
+    } else if (isObject(schema[key])) {
+      if (!schema[key].hasOwnProperty('type') || typeof schema[key].type !== 'string') {
+        extracted[key] = schema[key];
+        delete schema[key];
+        delete ogSchema[key];
+      }
+    }
+  }
+  return extracted;
+}
+export interface RelationType {
   type: 'Relation';
   model: string;
   required?: boolean;
